restructured UserGroups API fns
