+++ src/DataServices/DataServices.ts
@@ -1,7 +1,6 @@
 import { forkJoin, from, of, throwError } from 'rxjs';
 import { catchError, flatMap, map, tap } from 'rxjs/operators';
 import { Token } from '../Auth/types';
-import { UserGroupsData } from '../UserGroups/types';
 import { dataObjectToArray } from '../utils/Utils';
 import { DataSource, Options, User } from './types';
 
@@ -155,50 +154,6 @@ const fetchTimeseriesFullNames = (dataSources: DataSource[], token: string, grou
   return forkJoin(requests).pipe(map((ts) => ts.flat()));
 };
 
-// ACCOUNTS
-// Could be an account name or `me`.
-const fetchUserGroups = (host: string, token: string) =>
-  fetchUrl(`${host}/api/usergroups`, {
-    method: 'GET',
-    additionalHeaders: { Authorization: `Bearer ${token}` },
-  }).pipe(tap((res) => console.log('fetchUserGroups', res)));
-
-const updateUserGroupsForUser = (host: string, token: string, data: { userId: string; groups: string[] }) =>
-  fetchUrl(`${host}/api/usergroups/user/${data.userId}`, {
-    method: 'POST',
-    additionalHeaders: { Authorization: `Bearer ${token}` },
-    body: JSON.stringify(data.groups),
-  }).pipe(tap((res) => console.log('fetchUserGroups', res)));
-
-const createUserGroup = (host: string, token: string, data: UserGroupsData) =>
-  fetchUrl(`${host}/api/usergroups`, {
-    method: 'POST',
-    additionalHeaders: { Authorization: `Bearer ${token}` },
-    body: JSON.stringify(data),
-  }).pipe(tap((res) => console.log('createUserGroup', res)));
-
-const updateUserGroups = (host: string, token: string, data: UserGroupsData) =>
-  fetchUrl(`${host}/api/usergroups`, {
-    method: 'PUT',
-    additionalHeaders: { Authorization: `Bearer ${token}` },
-    body: JSON.stringify({
-      id: data.id,
-      name: data.name,
-      users: data.users,
-      metadata: data.metadata,
-    }),
-  }).pipe(tap((res) => console.log('updateUserGroups', res)));
-
-const deleteUserGroup = (host: string, token: string, id: string) =>
-  fetchUrl(`${host}/api/usergroups/${id}`, {
-    method: 'DELETE',
-    additionalHeaders: {
-      Authorization: `Bearer ${token}`,
-    },
-  }).pipe(tap((res) => console.log('deleted account', res)));
-
-// MAIL TEMPLATES
-
 // MAP
 
 const fetchMapAnimationFiles = (
@@ -329,11 +284,6 @@ export {
   validateToken,
   fetchUrl,
   fetchTimeseriesValues,
-  fetchUserGroups,
-  updateUserGroupsForUser,
-  createUserGroup,
-  updateUserGroups,
-  deleteUserGroup,
   fetchFeatureCollectionValues,
   fetchTimeseriesByGroup,
   fetchTimeseriesFullNames,
